backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
opencv-python==4.8.0.76
insightface==0.7.3
onnxruntime-gpu==1.16.3
scipy==1.10.1
